Refactor dynamo distributed test helpers to be reusable
[ghstack-poisoned]
